[refactor] #21 Date Format extension으로 분리
